тесты готовы

+++ src/main/DescendingComparator.java
@@ -4,11 +4,12 @@ import java.util.Comparator;
 
 
 public class DescendingComparator implements Comparator<Integer> {
+
     @Override
     public int compare(Integer value1, Integer value2) {
-        if(value1.intValue() < value2.intValue()) {
+        if(value1 < value2) {
             return -1;
-        } else if(value1.intValue() > value2.intValue()){
+        } else if (value1 > value2) {
             return 1;
         } else {
             return 0;
